Pictures added to the infoWindow and resizes based on size of screen.
However, I need help on the following issue: when the screen resizes (goes to small hamburger view or moves out from small screen hamburger view to full screen) the switch resets to list position. It does not remember what switch state it was in before. This means switching screen size can fool our switch to state we are in listview when we are actually in map view.

I'm not sure how I can fix this, unless resizing dispatches an action that is always false (list)